Changed a lot of authorization for the order but still working on improving it

+++ src/main/java/nl/novi/finalAssignmentBackend/Service/OrderService.java
@@ -6,6 +6,7 @@ import nl.novi.finalAssignmentBackend.Repository.ShoppingListRepository;
 import nl.novi.finalAssignmentBackend.entities.Order;
 import nl.novi.finalAssignmentBackend.entities.ShoppingList;
 import nl.novi.finalAssignmentBackend.exceptions.RecordNotFoundException;
+import nl.novi.finalAssignmentBackend.helper.LoggedInCheck;
 import nl.novi.finalAssignmentBackend.helper.OrderHelpers;
 import nl.novi.finalAssignmentBackend.helper.PDFCreator.PdfFileOrder;
 import nl.novi.finalAssignmentBackend.mappers.OrderMapper.OrderMapper;
@@ -24,61 +25,102 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final OrderMapper orderMapper;
     private final ShoppingListRepository shoppingListRepository;
-
     private final OrderHelpers orderHelpers;
+    private final LoggedInCheck loggedInCheck;
 
-    public OrderService(OrderRepository orderRepository, OrderMapper orderMapper, ShoppingListRepository shoppingListRepository, OrderHelpers orderHelpers) {
+    public OrderService(OrderRepository orderRepository, OrderMapper orderMapper, ShoppingListRepository shoppingListRepository, OrderHelpers orderHelpers, LoggedInCheck loggedInCheck) {
         this.orderRepository = orderRepository;
         this.orderMapper = orderMapper;
         this.shoppingListRepository = shoppingListRepository;
         this.orderHelpers = orderHelpers;
+        this.loggedInCheck = loggedInCheck;
     }
 
-//    public List<OrderModel> getOrders(){
-//        return orderRepository.findAll().stream().map(orderMapper::fromEntity).collect(Collectors.toList());
+//    public List<OrderModel>getAllOrders(String username) {
+//        loggedInCheck.verifyLoggedInUser(username);
+//
+//        List<Order>orders=orderRepository.findAll();
+//        for(Order order: orders){
+//            loggedInCheck.verifyOwnerAuthorization(order.getUser().getUsername(), username, "order" );
+//            if (order.getUser().getAuthorities().contains("ADMIN")){
+//                orders.add(order);
+//            }
+//            if (order.getUser().getAuthorities().equals("USER")){
+//                orders.add(order);
+//            }
+//            if (order.getUser().getUsername().isEmpty()){
+//                throw new EntityNotFoundException("Orders cannot excist without users, please add a user");
+//        }
+
+//
+//        {
+//
+//            List<Order>orderWithUsers = orders.stream()
+//                .filter(order -> order.getUser() != null)
+//                .collect(Collectors.toList());
+//
+//        }
+//        return orderWithUsers.stream()
+//                .map(orderMapper::fromEntity)
+//                .collect(Collectors.toList());
+//        }
+//        else if (username.equals("USER")){
+//            List<Order>orders=orderRepository.findOrderByUser(username);
+//            if (orders.isEmpty()){
+//                throw new EntityNotFoundException("No orders found with username " + username);
+//            }
+//            return orders.stream()
+//                    .map(orderMapper::fromEntity)
+//                    .collect(Collectors.toList());
+//        }
+//        throw new EntityNotFoundException("no orders found");
 //    }
+        public List<OrderModel> getAllOrders() {
 
-    public List<OrderModel>getAllOrders(){
-        List<Order>orders = orderRepository.findAll();
+            List<Order> orders = orderRepository.findAll();
 
-        List<Order>orderWithUsers = orders.stream()
-                .filter(order -> order.getUser() != null)
-                .collect(Collectors.toList());
-        if (orderWithUsers.isEmpty()){
-            throw new EntityNotFoundException("Orders cannot excist without users, please add a user");
+            List<Order> orderWithUsers = orders.stream()
+                    .filter(order -> order.getUser() != null)
+                    .collect(Collectors.toList());
+            if (orderWithUsers.isEmpty()) {
+                throw new EntityNotFoundException("Orders cannot excist without users, please add a user");
+            }
+
+            return orderWithUsers.stream()
+                    .map(orderMapper::fromEntity)
+                    .collect(Collectors.toList());
         }
 
-        return orderWithUsers.stream()
-                .map(orderMapper::fromEntity)
-                .collect(Collectors.toList());
-    }
 
-    public OrderModel getOrderById(Long id){
-       Order order= orderRepository.findById(id).orElseThrow(()-> new EntityNotFoundException("Order not found with id " + id));
+    public OrderModel getOrderById(Long id, String username){
+        loggedInCheck.verifyLoggedInUser(username);
+
+        Order order= orderRepository.findById(id).orElseThrow(()-> new EntityNotFoundException("Order not found with id " + id));
         if(order.getUser() == null){
             throw new EntityNotFoundException("please assign a user to the order");
         }
-       orderHelpers.calculateTotalPrice(id);
-       orderRepository.save(order);
-       return orderMapper.fromEntity(order);
+        loggedInCheck.verifyOwnerAuthorization(order.getUser().getUsername(), username, "order" );
+
+        orderHelpers.calculateTotalPrice(id);
+        orderRepository.save(order);
+        return orderMapper.fromEntity(order);
     }
 
-    public OrderModel createOrder(OrderModel orderModel){
-//        Order order = orderMapper.toEntity(orderModel);
-//        order = orderRepository.save(order);
-//        return orderMapper.fromEntity(order);
+    public OrderModel createOrder(){
         Order order = new Order();
-        order.setCreatePdf(orderModel.isCreatePdf());
-        order.setOrderConfirmation(orderModel.getOrderConfirmation());
-
         order = orderRepository.save(order);
         return orderMapper.fromEntity(order);
     }
 
-    public OrderModel updateOrder(Long id, OrderModel orderModel){
+    public OrderModel updateOrder(Long id, String username, OrderModel orderModel){
+
+        loggedInCheck.verifyLoggedInUser(username);
+
         Optional<Order> orderFound = orderRepository.findById(id);
+
         if(orderFound.isPresent()){
             Order excistingOrder = orderFound.get();
+            loggedInCheck.verifyOwnerAuthorization(excistingOrder.getUser().getUsername(), username, "order" );
 
             if(orderModel.getOrderConfirmation() != null){
                 excistingOrder.setOrderConfirmation(orderModel.getOrderConfirmation());
@@ -91,7 +133,7 @@ public class OrderService {
                 excistingOrder.setCreatePdf(orderModel.isCreatePdf());
             }
             if (orderModel.getOrderConfirmation() && orderModel.isCreatePdf()){
-                createOrder(excistingOrder);
+                createPdfOfOrder(excistingOrder);
             }
             excistingOrder = orderRepository.save(excistingOrder);
             return orderMapper.fromEntity(excistingOrder);
@@ -100,9 +142,14 @@ public class OrderService {
         }
     }
 
-    public void addShoppingListToOrder(Long orderId, Long shoppingListId){
+    public void addShoppingListToOrder(Long orderId,Long shoppingListId, String username){
+
+        loggedInCheck.verifyLoggedInUser(username);
+
         Order order = orderRepository.findById(orderId).orElseThrow(()-> new EntityNotFoundException("Order with id " + orderId + " not found"));
         ShoppingList shoppingList = shoppingListRepository.findById(shoppingListId).orElseThrow(()-> new EntityNotFoundException("Shopping list with id " + shoppingListId + " not found"));
+        loggedInCheck.verifyOwnerAuthorization(shoppingList.getUser().getUsername(), username, "order" );
+
         if(Objects.equals(shoppingList.getType(), "wishlist") || !Objects.equals(shoppingList.getType(), "shoppinglist")){
             throw new EntityNotFoundException("please change the type of the wishlist with id " + shoppingListId + " to shoppingList before adding it to the order");
         }
@@ -117,17 +164,26 @@ public class OrderService {
         }
         if(shoppingList.getGames().isEmpty() || shoppingList.getMovies().isEmpty()){
             throw new EntityNotFoundException("the shoppinglist cannot be added to the order when empty, please add a game or movie to the shoppingList before adding it to the order!");
-        }// when the orderConfirmation is true this doesnt work, how to prevent the orderConfirmation to be set to true before a shoppingList has been added upon creation of a new order?
+        }
         order.getShoppingList().add(shoppingList);
         orderRepository.save(order);
 
     }
 
-    public void deleteOrder(Long id){
+    public void deleteOrder(Long id, String username){
+        loggedInCheck.verifyLoggedInUser(username);
+
+        Optional<Order>optionalOrder = orderRepository.findById(id);
+        if(optionalOrder.isEmpty()){
+            throw new RecordNotFoundException("Order with id " + id + "does not exist!");
+        }
+        Order order = optionalOrder.get();
+        loggedInCheck.verifyOwnerAuthorization(order.getUser().getUsername(),username,"order");
+
         orderRepository.deleteById(id);
     }
 
-    public void createOrder(Order order) {
+    public void createPdfOfOrder(Order order) {
         if (order.isCreatePdf()) {
             PdfFileOrder pdfFileOrder = new PdfFileOrder();
             try {
